Added docker attempt
