+++ docker-compose.yml
@@ -1,0 +1,19 @@
+version: '3'
+
+services:
+  peer1:
+    image: peer
+    ports:
+      - "8081:8080"
+    networks:
+      - peer-network
+
+  peer2:
+    image: peer
+    ports:
+      - "8082:8080"
+    networks:
+      - peer-network
+
+networks:
+  peer-network:
